requirements_complete.txt
# Multi-Agent ML Pipeline - Complete Dependencies
# Combined from all agent implementations

# ============================================================================
# CORE DATA SCIENCE LIBRARIES
# ============================================================================
pandas>=2.1.4
numpy>=1.24.3
scipy>=1.10.0
scikit-learn>=1.3.2
joblib>=1.3.2

# ============================================================================
# MACHINE LEARNING LIBRARIES
# ============================================================================
lightgbm>=4.1.0
xgboost>=2.0.3
catboost>=1.2.0

# ============================================================================
# VISUALIZATION LIBRARIES
# ============================================================================
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0

# ============================================================================
# LANGCHAIN AND LLM INTEGRATION
# ============================================================================
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.1.0
langgraph>=0.1.15

# ============================================================================
# LLM PROVIDERS
# ============================================================================
ollama>=0.3.3
openai>=1.12.0

# ============================================================================
# SLACK INTEGRATION
# ============================================================================
slack-bolt>=1.18.1
slack-sdk>=3.28.0

# ============================================================================
# DATA VALIDATION AND MODELS
# ============================================================================
pydantic>=2.0.0
typing-extensions>=4.0.0

# ============================================================================
# FILE HANDLING AND I/O
# ============================================================================
requests>=2.31.0
openpyxl>=3.1.0      # Excel file support
xlrd>=2.0.1          # Legacy Excel file support
python-dotenv>=1.0.0 # Environment variables

# ============================================================================
# STATISTICAL ANALYSIS
# ============================================================================
statsmodels>=0.14.0
scipy>=1.11.0

# ============================================================================
# PERFORMANCE AND OPTIMIZATION
# ============================================================================
numba>=0.57.0        # Optional: For enhanced performance
tqdm>=4.65.0         # Progress bars

# ============================================================================
# DATABASE AND PERSISTENCE
# ============================================================================
sqlite3              # Built-in Python module for LangGraph persistence

# ============================================================================
# DEVELOPMENT AND TESTING (Optional)
# ============================================================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.7.0
# flake8>=6.0.0
# jupyter>=1.0.0

# ============================================================================
# SYSTEM AND UTILITIES
# ============================================================================
psutil>=5.9.0        # System monitoring
pathlib              # Built-in Python module
tempfile             # Built-in Python module
datetime             # Built-in Python module
json                 # Built-in Python module
os                   # Built-in Python module
sys                  # Built-in Python module
traceback            # Built-in Python module
warnings             # Built-in Python module
logging              # Built-in Python module
re                   # Built-in Python module
time                 # Built-in Python module
collections          # Built-in Python module
dataclasses          # Built-in Python module
enum                 # Built-in Python module
functools            # Built-in Python module
itertools            # Built-in Python module
pickle               # Built-in Python module
shutil               # Built-in Python module
subprocess           # Built-in Python module

# ============================================================================
# INSTALLATION INSTRUCTIONS
# ============================================================================
# 
# To install all dependencies, run:
# pip install -r requirements_complete.txt
#
# For minimal installation (core functionality only):
# pip install pandas numpy scikit-learn pydantic
#
# For Slack integration:
# pip install slack-bolt slack-sdk
#
# For LLM integration:
# pip install langchain langchain-core langchain-openai langchain-ollama langgraph
#
# For local LLM support:
# pip install ollama
#
# For OpenAI support:
# pip install openai
#
